example2/automatic/service1/requirements.txt
Flask==3.0.0
requests==2.31.0
opentelemetry-api>=1.20.0,<2.0
opentelemetry-sdk>=1.20.0,<2.0
opentelemetry-exporter-jaeger-thrift>=1.20.0,<2.0
opentelemetry-instrumentation-flask>=0.41b0,<1.0
opentelemetry-instrumentation-requests>=0.41b0,<1.0
Deprecated>=1.2.13
